+++ catalyst-gateway/bin/src/service/docs/mod.rs
@@ -8,13 +8,11 @@ use rust_embed::RustEmbed;
 pub(crate) fn docs<T: OpenApi + 'static, W: Webhook + 'static>(
     api_service: &OpenApiService<T, W>,
 ) -> Route {
-    let spec = api_service.spec();
-
     let swagger_ui = api_service.swagger_ui();
     let rapidoc_ui = api_service.rapidoc();
     let redoc_ui = api_service.redoc();
     let openapi_explorer = api_service.openapi_explorer();
-    let stoplight_ui = stoplight_elements::create_endpoint(&spec);
+    let stoplight_ui = stoplight_elements::create_endpoint(&api_service.spec());
 
     Route::new()
         .at("/", get(stoplight_ui))
@@ -22,10 +20,8 @@ pub(crate) fn docs<T: OpenApi + 'static, W: Webhook + 'static>(
         .nest("/redoc", redoc_ui)
         .nest("/rapidoc", rapidoc_ui)
         .nest("/openapi_explorer", openapi_explorer)
-        .at(
-            "/cat-gateway.json",
-            poem::endpoint::make_sync(move |_| spec.clone()),
-        )
+        .at("/cat-gateway.json", api_service.spec_endpoint())
+        .at("/cat-gateway.yml", api_service.spec_endpoint_yaml())
 }
 
 /// Embed static files.
