+++ catalyst-gateway/bin/src/service_settings.rs
@@ -41,7 +41,7 @@ const API_URL_PREFIX_DEFAULT: &str = "/api";
 /// the URL to the `PostgreSQL` event database,
 /// and the logging level.
 #[derive(Args, Clone)]
-pub(crate) struct Settings {
+pub(crate) struct ServiceSettings {
     /// Server binding address
     #[clap(long, default_value = ADDRESS_DEFAULT)]
     pub(crate) address: SocketAddr,
@@ -256,10 +256,10 @@ mod tests {
     // }
 
     #[test]
-    fn generate_github_issue_url() {
+    fn generate_github_issue_url_test() {
         let title = "Hello, World! How are you?";
         assert_eq!(
-            super::generate_github_issue_url(title).unwrap().as_str(),
+            generate_github_issue_url(title).unwrap().as_str(),
             "https://github.com/input-output-hk/catalyst-core/issues/new?template=bug_report.md&title=Hello%2C+World%21+How+are+you%3F"
         );
     }
